IACreatorsChat/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy==2.0.32
aiosqlite==0.20.0
httpx==0.27.0
